split data models
