+++ tests/unit/test_wrroc_models.py
@@ -1,8 +1,14 @@
 import unittest
 from pydantic import ValidationError
 
-from crategen.models import WRROCProcess, WRROCWorkflow, WRROCProvenance, WRROCDataWES
+from crategen.models.wrroc_models import (
+    WRROCProcess,
+    WRROCWorkflow,
+    WRROCProvenance,
+    WRROCDataWES,
+)
 from crategen.validators import validate_wrroc, validate_wrroc_tes, validate_wrroc_wes
+
 
 class TestWRROCProcessModel(unittest.TestCase):
     """
@@ -19,7 +25,12 @@ class TestWRROCProcessModel(unittest.TestCase):
             "description": "A simple process",
             "startTime": "2024-07-10T14:30:00Z",
             "endTime": "2024-07-10T15:30:00Z",
-            "object": [{"id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/README.md", "name": "Input 1"}]
+            "object": [
+                {
+                    "id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/README.md",
+                    "name": "Input 1",
+                }
+            ],
         }
         model = WRROCProcess(**data)
         self.assertEqual(model.id, "process-id")
@@ -29,11 +40,7 @@ class TestWRROCProcessModel(unittest.TestCase):
         """
         Test that the WRROCProcess model handles empty object lists correctly.
         """
-        data = {
-            "id": "process-id",
-            "name": "Test Process",
-            "object": []
-        }
+        data = {"id": "process-id", "name": "Test Process", "object": []}
         model = WRROCProcess(**data)
         self.assertEqual(model.object, [])
 
@@ -43,10 +50,11 @@ class TestWRROCProcessModel(unittest.TestCase):
         """
         data = {
             "id": 123,  # id should be a string
-            "name": None  # name should be a string
+            "name": None,  # name should be a string
         }
         with self.assertRaises(ValidationError):
             WRROCProcess(**data)
+
 
 class TestWRROCWorkflowModel(unittest.TestCase):
     """
@@ -62,23 +70,26 @@ class TestWRROCWorkflowModel(unittest.TestCase):
             "name": "Test Workflow",
             "workflowType": "CWL",
             "workflowVersion": "v1.0",
-            "result": [{"id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE", "name": "Output 1"}]
+            "result": [
+                {
+                    "id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE",
+                    "name": "Output 1",
+                }
+            ],
         }
         model = WRROCWorkflow(**data)
         self.assertEqual(model.workflowType, "CWL")
-        self.assertEqual(model.result[0]['name'], "Output 1")
+        self.assertEqual(model.result[0]["name"], "Output 1")
 
     def test_wrroc_workflow_missing_optional_fields(self):
         """
         Test that the WRROCWorkflow model handles missing optional fields correctly.
         """
-        data = {
-            "id": "workflow-id",
-            "name": "Test Workflow"
-        }
+        data = {"id": "workflow-id", "name": "Test Workflow"}
         model = WRROCWorkflow(**data)
         self.assertIsNone(model.workflowType)
         self.assertIsNone(model.workflowVersion)
+
 
 class TestWRROCProvenanceModel(unittest.TestCase):
     """
@@ -93,23 +104,20 @@ class TestWRROCProvenanceModel(unittest.TestCase):
             "id": "provenance-id",
             "name": "Test Provenance",
             "provenanceData": "Provenance information",
-            "agents": [{"id": "agent1", "name": "Agent 1"}]
+            "agents": [{"id": "agent1", "name": "Agent 1"}],
         }
         model = WRROCProvenance(**data)
         self.assertEqual(model.provenanceData, "Provenance information")
-        self.assertEqual(model.agents[0]['name'], "Agent 1")
+        self.assertEqual(model.agents[0]["name"], "Agent 1")
 
     def test_wrroc_provenance_empty_agents_list(self):
         """
         Test that the WRROCProvenance model handles empty agents lists correctly.
         """
-        data = {
-            "id": "provenance-id",
-            "name": "Test Provenance",
-            "agents": []
-        }
+        data = {"id": "provenance-id", "name": "Test Provenance", "agents": []}
         model = WRROCProvenance(**data)
         self.assertEqual(model.agents, [])
+
 
 class TestWRROCValidators(unittest.TestCase):
     """
@@ -120,10 +128,7 @@ class TestWRROCValidators(unittest.TestCase):
         """
         Test that validate_wrroc correctly identifies a WRROCProcess entity.
         """
-        data = {
-            "id": "process-id",
-            "name": "Test Process"
-        }
+        data = {"id": "process-id", "name": "Test Process"}
         model = validate_wrroc(data)
         self.assertIsInstance(model, WRROCProcess)
 
@@ -135,7 +140,7 @@ class TestWRROCValidators(unittest.TestCase):
             "id": "workflow-id",
             "name": "Test Workflow",
             "workflowType": "CWL",
-            "workflowVersion": "v1.0"
+            "workflowVersion": "v1.0",
         }
         model = validate_wrroc(data)
         self.assertIsInstance(model, WRROCWorkflow)
@@ -147,7 +152,7 @@ class TestWRROCValidators(unittest.TestCase):
         data = {
             "id": "provenance-id",
             "name": "Test Provenance",
-            "provenanceData": "Provenance information"
+            "provenanceData": "Provenance information",
         }
         model = validate_wrroc(data)
         self.assertIsInstance(model, WRROCProvenance)
@@ -156,9 +161,7 @@ class TestWRROCValidators(unittest.TestCase):
         """
         Test that validate_wrroc raises a ValueError for invalid WRROC data.
         """
-        data = {
-            "unknown_field": "unexpected"
-        }
+        data = {"unknown_field": "unexpected"}
         with self.assertRaises(ValueError):
             validate_wrroc(data)
 
@@ -169,8 +172,18 @@ class TestWRROCValidators(unittest.TestCase):
         data = {
             "id": "process-id",
             "name": "Test Process",
-            "object": [{"id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/README.md", "name": "Input 1"}],
-            "result": [{"id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE", "name": "Output 1"}]
+            "object": [
+                {
+                    "id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/README.md",
+                    "name": "Input 1",
+                }
+            ],
+            "result": [
+                {
+                    "id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE",
+                    "name": "Output 1",
+                }
+            ],
         }
         model = validate_wrroc_tes(data)
         self.assertEqual(model.id, "process-id")
@@ -184,7 +197,12 @@ class TestWRROCValidators(unittest.TestCase):
             "id": "process-id",
             "name": "Test Process",
             "object": [],
-            "result": [{"id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE", "name": "Output 1"}]
+            "result": [
+                {
+                    "id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE",
+                    "name": "Output 1",
+                }
+            ],
         }
         model = validate_wrroc_tes(data)
         self.assertEqual(model.object, [])
@@ -193,29 +211,35 @@ class TestWRROCValidators(unittest.TestCase):
         """
         Test that validate_wrroc_tes raises a ValueError if required fields for TES conversion are missing.
         """
-        data = {
-            "id": "process-id",
-            "name": "Test Process"
-        }
+        data = {"id": "process-id", "name": "Test Process"}
         with self.assertRaises(ValueError):
             validate_wrroc_tes(data)
 
     def test_validate_wrroc_wes(self):
-            """
-            Test that validate_wrroc_wes correctly validates a WRROC entity for WES conversion.
-            """
-            data = {
-                "id": "workflow-id",
-                "name": "Test Workflow",
-                "workflowType": "CWL",
-                "workflowVersion": "v1.0",
-                "status": "completed", 
-                "object": [{"id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/README.md", "name": "Input 1"}],
-                "result": [{"id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE", "name": "Output 1"}]
-            }
-            model = validate_wrroc_wes(data)
-            self.assertIsInstance(model, WRROCDataWES)
-
+        """
+        Test that validate_wrroc_wes correctly validates a WRROC entity for WES conversion.
+        """
+        data = {
+            "id": "workflow-id",
+            "name": "Test Workflow",
+            "workflowType": "CWL",
+            "workflowVersion": "v1.0",
+            "status": "completed",
+            "object": [
+                {
+                    "id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/README.md",
+                    "name": "Input 1",
+                }
+            ],
+            "result": [
+                {
+                    "id": "https://raw.githubusercontent.com/elixir-cloud-aai/CrateGen/main/LICENSE",
+                    "name": "Output 1",
+                }
+            ],
+        }
+        model = validate_wrroc_wes(data)
+        self.assertIsInstance(model, WRROCDataWES)
 
     def test_validate_wrroc_wes_invalid_url(self):
         """
@@ -234,7 +258,7 @@ class TestWRROCValidators(unittest.TestCase):
                 "name": "Test Workflow",
                 "workflowType": "CWL",
                 "workflowVersion": "v1.0",
-                "result": [{"id": url, "name": "Output 1"}]
+                "result": [{"id": url, "name": "Output 1"}],
             }
             with self.assertRaises(ValueError):
                 validate_wrroc_wes(data)
@@ -243,12 +267,10 @@ class TestWRROCValidators(unittest.TestCase):
         """
         Test that validate_wrroc_wes raises a ValueError if required fields for WES conversion are missing.
         """
-        data = {
-            "id": "workflow-id",
-            "name": "Test Workflow"
-        }
+        data = {"id": "workflow-id", "name": "Test Workflow"}
         with self.assertRaises(ValueError):
             validate_wrroc_wes(data)
+
 
 if __name__ == "__main__":
     unittest.main()
